+++ changelog.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Added binary distribution for linux aarch64 (Streamlit's environment)
+- Added new separator option in eds.table and new input check
 
 ### Changed